+++ front_end/src/components/cards/account/AccountLikesCard.js
@@ -20,7 +20,7 @@ import IncreaseIcon from '@material-ui/icons/ArrowDropUp'
 import DecreaseIcon from '@material-ui/icons/ArrowDropDown'
 import AssistantIcon from '@material-ui/icons/Assistant'
 import ThumbsUpIcon from '@material-ui/icons/ThumbUp'
-import BasicChart from '../../diagrams/chartTest'
+import BasicChart from '../../diagrams/BasicChart'
 import green from "@material-ui/core/es/colors/green";
 
 const styles = theme => ({
